file separating and little refactoring dynamicstruct
