+++ dynamicstruct/builder_test.go
